refactor(gsm_arena): separate utilities
