+++ src/korone/modules/gsm_arena/utils/__init__.py
@@ -1,0 +1,2 @@
+# SPDX-License-Identifier: BSD-3-Clause
+# Copyright (c) 2024 Hitalo M. <https://github.com/HitaloM>
